+++ tomcat/src/main/java/nextstep/jwp/handler/Handler.java
@@ -1,12 +1,14 @@
 package nextstep.jwp.handler;
 
 import java.io.IOException;
+import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.HttpResponseParser;
 
 public class Handler {
 
-    private static final RequestHandler requestHandler = new RequestHandler();
+    private static final RequestHandler requestHandler = new RequestHandler(new SessionManager());
 
     public static String handle(HttpRequest request) {
         try {
