+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.net.Socket;
 import nextstep.jwp.exception.UncheckedServletException;
 import nextstep.jwp.handler.Handler;
-import nextstep.jwp.handler.HttpRequestParser;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
